Proposal Form Height Growth
Corrected proposal form height growth. On resizing the proposal form, can resize would re-calc the number of lines shown. The issue with this is that the height of the boarder was not being removed before calculation. This would mean the height would be increased if the boarder width could accommodate another  item. The growth would appear in the next show, and cause another resize. This would repeat.