Reduce Flicker
Removed unnecessary invalidations
UpdateScrollbars calls Invalidate causing full repaint.
As it stands moving the cursor eg. left or rigth and typing one
character results in a full repaint and flicker.

Three changes were made that resolve the issue.
a) In SetCaretXYEx I removed the line
Include(fStateFlags, sfScrollbarChanged);
since it is not needed.  Scrollbars are updated when TopLine or LeftChar
get changed.
b)   I removed UpdateScrollbars for LinesChanged and added
Include(fStateFlags, sfScrollbarChanged);
to ListInserted, ListDeleted.  Thus when the user types code in one line
this does not result in a full repaint.
b) In LinesInserted I changed the line
UnionRect(fInvalidateRect, fInvalidateRect, rcInval)
to
UnionRect(fInvalidateRect, rcInval, fInvalidateRect)
This subtle change prevents the invalidate of the lines above the line
you are typing when fInvalidateRect is empty.